rename to DocsSearch; issues
